+++ app/src/main/java/com/example/streamtv/Profile/ProfileFragment.kt
@@ -1,4 +1,4 @@
-package com.example.streamtv
+package com.example.streamtv.Profile
 
 import android.app.AlertDialog
 import android.net.Uri
@@ -12,6 +12,8 @@ import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.NavHostFragment
 import com.bumptech.glide.Glide
+import com.example.streamtv.R
+import com.example.streamtv.Utils.User
 import com.example.streamtv.databinding.ProfileBinding
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.auth.FirebaseUser
