+++ app/src/main/java/com/example/streamtv/Signup/SignUpFragment.kt
@@ -1,4 +1,4 @@
-package com.example.streamtv
+package com.example.streamtv.Signup
 
 import android.os.Bundle
 import android.util.Log
@@ -10,6 +10,9 @@ import android.widget.Toast
 import androidx.fragment.app.Fragment
 import androidx.navigation.NavController
 import androidx.navigation.fragment.NavHostFragment
+import com.example.streamtv.R
+import com.example.streamtv.Utils.User
+import com.example.streamtv.Utils.Validator
 import com.example.streamtv.databinding.SignUpBinding
 import com.google.firebase.auth.FirebaseAuth
 import com.google.firebase.database.FirebaseDatabase
