+++ app/src/main/java/com/example/streamtv/Utils/User.kt
@@ -1,4 +1,4 @@
-package com.example.streamtv
+package com.example.streamtv.Utils
 
 import com.google.firebase.database.IgnoreExtraProperties
