+++ app/src/main/res/layout/edit_profile.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".EditProfileFragment"
+        tools:context=".EditProfile.EditProfileFragment"
         android:orientation="vertical"
         android:paddingStart="24dp"
         android:paddingEnd="24dp">
