+++ app/src/main/res/layout/sign_up.xml
@@ -4,7 +4,7 @@
         xmlns:app="http://schemas.android.com/apk/res-auto"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".SignUpFragment">
+        tools:context=".Signup.SignUpFragment">
 
     <ImageView
             android:id="@+id/logo"
